refactor: move handler to internal
