+++ internal/handler/register_handler_test.go
